Feat: add authentication for dashboard file upload
